Fixes small bug in reading O3 excel file format #2 (there are two version of the O3 audit form).
The python script has to know how many rows to skip in order to read in an excel file. I changed the number of rows to skip from 24 to 25 on line 259 of this file (line 258 in 0.0.1). This version of the form was rarer, so I wasn't able to test it as robustly before.